+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/dispatcherservlet/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc.dispatcherservlet;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
